+++ packages/@guardian/source-react-components-development-kitchen/src/index.ts
@@ -30,8 +30,8 @@ export type { ErrorSummaryProps } from './summary/ErrorSummary';
 export type { SuccessSummaryProps } from './summary/SuccessSummary';
 export type { InfoSummaryProps } from './summary/InfoSummary';
 
-export { ToggleSwitch } from './toggle-switch/ToggleSwitch';
-export type { ToggleSwitchProps } from './toggle-switch/ToggleSwitch';
+export { ToggleSwitchApps } from './toggle-switch-apps/ToggleSwitchApps';
+export type { ToggleSwitchAppsProps } from './toggle-switch-apps/ToggleSwitchApps';
 
 export { FooterWithContents } from './footer-with-contents/FooterWithContents';
 export {
